pulsar-client-python-150/requirements.txt
celery==5.1.2
pulsar-client[avro]==2.10.2
